+++ src/host.rs
@@ -1,10 +1,10 @@
-use crate::messages::Message;
+use crate::messages::{Message, UiEvents, ParticipantStatus};
 
 use message_io::network::Endpoint;
 
 use bimap::BiMap;
 
-use message_io::events::{EventQueue};
+use message_io::events::{EventQueue, EventSender};
 use message_io::network::{Network, NetEvent, Transport};
 
 use hlua::{Lua, AnyLuaValue, LuaTable};
@@ -13,20 +13,14 @@ use std::io::Read;
 use crate::lua::SerdeLuaTable;
 use std::time::Duration;
 
-enum Event {
-    Network(NetEvent<Message>),
-    SendCode(String),
-    SendData,
-    Pause(Endpoint),
-    Play(Endpoint),
-    Stop(Endpoint),
-    Execute
-}
+use crate::messages::HostEvent;
 
 pub struct Host<'a> {
     participants: BiMap<String, Endpoint>,
-    event_queue: EventQueue<Event>,
+    event_queue: EventQueue<HostEvent>,
     network: Network,
+
+    ui_sender: Option<EventSender<UiEvents>>,
 
     participants_finished: usize,
     participants_startedwith: BiMap<String, Endpoint>,
@@ -36,13 +30,13 @@ pub struct Host<'a> {
 
 impl<'a> Host<'a> {
 
-    pub fn new(server_address: &str) -> Self {
+    pub fn new(server_address: &str) -> Result<Self, String> {
 
         let mut event_queue = EventQueue::new();
 
         let network_sender = event_queue.sender().clone();
 
-        let mut network = Network::new(move |net_event| network_sender.send(Event::Network(net_event)));
+        let mut network = Network::new(move |net_event| network_sender.send(HostEvent::Network(net_event)));
 
         let mut lua = Lua::new();
 
@@ -50,40 +44,33 @@ impl<'a> Host<'a> {
 
         match network.listen(Transport::Tcp, server_address) {
             Ok(_) => println!("TCP Server running at {}", server_address),
-            Err(e) => panic!("Can not listen at {} - {}", server_address, e)
+            Err(e) => return Err(format!("Can not listen at {} - {}", server_address, e))
         }
 
-        Host {
+
+        Ok(Host {
             participants: BiMap::new(),
             event_queue,
             network,
             participants_finished: 0,
             participants_startedwith: BiMap::new(),
+            ui_sender: None,
             lua
-        }
+        })
     }
 
-    /*pub fn add_code(& mut self, path: & str) {
-        // Get source code
+    pub fn get_host_sender(& mut self) -> EventSender<HostEvent> {
+        self.event_queue.sender().clone()
+    }
 
-        use std::fs::File;
-
-        let mut fh = File::open(path).unwrap();
-
-        let mut source_code = String::new();
-
-        fh.read_to_string(& mut source_code).unwrap();
-
-        match self.lua.execute::<()>(source_code.as_str()) {
-            Ok(_) => {}
-            Err(e) => { panic!("LuaError: {:?}", e); }
-        }
-
-
-    }*/
+    pub fn set_ui_sender(& mut self, sender: EventSender<UiEvents>) {
+        self.ui_sender = Some(sender);
+    }
 
     pub fn start_participants(& mut self, path: &str) {
 
+        self.ui_sender.as_ref().unwrap().send(UiEvents::HostMessage(format!("Starting calculations.")));
+
         use std::fs::File;
 
         let mut fh = File::open(path).unwrap();
@@ -94,60 +81,67 @@ impl<'a> Host<'a> {
 
         match self.lua.execute::<()>(source_code.as_str()) {
             Ok(_) => {}
-            Err(e) => { panic!("LuaError: {:?}", e); }
+            Err(e) => { panic!("DEPRECIATED LuaError: {:?}", e); }
         }
+
+        self.participants_finished = 0;
 
         self.participants_startedwith = self.participants.clone();
 
-        self.event_queue.sender().send(Event::SendData);
+        self.event_queue.sender().send(HostEvent::SendData);
 
-        self.event_queue.sender().send(Event::SendCode(source_code));
+        self.event_queue.sender().send(HostEvent::SendCode(source_code));
 
-        self.event_queue.sender().send(Event::Execute);
+        self.event_queue.sender().send(HostEvent::Execute);
     }
 
     pub fn display_participants(& self) {
-        println!("Participants: {:?}", self.participants);
+        println!("DEPRECIATED Participants: {:?}", self.participants);
     }
 
     pub fn display_participant_count(& self) {
-        println!("Participant count: {}", self.participants.len());
+        println!("DEPRECIATED Participant count: {}", self.participants.len());
     }
 
     pub fn check_events(& mut self) {
 
-        match self.event_queue.receive_timeout(Duration::from_micros(1)) {
+        match self.event_queue.receive_timeout(Duration::from_micros(0)) {
             Some(event) => match event {
-                Event::Network(net_event) => match net_event {
+                HostEvent::Network(net_event) => match net_event {
                     NetEvent::Message(endpoint, message) => {
 
 
                         match message {
                             Message::Register(name) => {
-                                println!("Register participant '{}'", name);
                                 if self.participants.contains_left(&name) {
-                                    println!("Participant {} could not be registered. Participant with this name already exists.", name);
+                                    self.ui_sender.as_ref().unwrap().send(UiEvents::HostMessage(format!("Participant {} could not be registered. Participant with this name already exists.", name)));
                                     self.network.remove_resource(endpoint.resource_id());
                                 }
                                 else {
-                                    self.participants.insert(name, endpoint);
+                                    self.participants.insert(name.clone(), endpoint);
+                                    self.ui_sender.as_ref().unwrap().send(UiEvents::ParticipantRegistered(endpoint, name.clone()));
+                                    self.ui_sender.as_ref().unwrap().send(UiEvents::ChangeStatusTo(ParticipantStatus::Idle, endpoint, name));
+
                                 }
                             },
                             Message::Unregister => {
-                                let endpoint_name = self.participants.get_by_right(&endpoint).unwrap();
-                                println!("Unregister participant '{}'", endpoint_name);
+                                {
+                                    let endpoint_name = self.participants.get_by_right(&endpoint).unwrap();
+                                    self.ui_sender.as_ref().unwrap().send(UiEvents::ParticipantUnregistered(endpoint, endpoint_name.clone()));
+                                }
                                 self.participants.remove_by_right(&endpoint);
                             },
                             Message::VectorPTH(data) => {
 
+                                let endpoint_name = self.participants.get_by_right(&endpoint).unwrap();
+
+                                self.ui_sender.as_ref().unwrap().send(UiEvents::ChangeStatusTo(ParticipantStatus::Idle, endpoint, endpoint_name.clone()));
+
                                 if self.participants_startedwith != self.participants {
-                                    println!("Some participants have disconnected/connected before execution could complete.")
+                                    self.ui_sender.as_ref().unwrap().send(UiEvents::HostMessage(format!("Some participants have disconnected/connected before execution could complete.")));
+
                                 }
                                 else {
-                                    let endpoint_name = self.participants.get_by_right(&endpoint).unwrap();
-
-                                    println!("Received data from '{}'.", endpoint_name);
-
                                     //A participant has finished, so increment the count
                                     self.participants_finished += 1;
 
@@ -169,11 +163,6 @@ impl<'a> Host<'a> {
                                     //Move the temporary table to to the global results
                                     self.lua.execute::<()>(format!("results[{}] = tmp_table", self.participants_finished).as_str()).unwrap();
 
-                                    /*println!("    len:      {}", self.participants.len());
-                                    println!("    finished: {}", self.participants_finished);
-                                    println!("    parts_:   {:?}", self.participants);
-                                    println!("    partsS:   {:?}", self.participants_startedwith);*/
-
                                     // Test to see if all participants have finished
                                     if self.participants.len() == self.participants_finished {
 
@@ -184,46 +173,51 @@ impl<'a> Host<'a> {
                                         // Get return value
                                         let return_code: String = interpret_results.call().unwrap();
 
-                                        println!("All participants finished. interpret_results return code: {}", return_code);
+                                        self.ui_sender.as_ref().unwrap().send(UiEvents::InterpretResultsReturn(return_code));
 
-
-                                        self.participants_finished = 0;
                                     }
                                 }
                             },
                             Message::ParticipantError(err) => {
                                 let endpoint_name = self.participants.get_by_right(&endpoint).unwrap();
-                                println!("Participant '{}' Error: {}", endpoint_name, err);
+                                self.ui_sender.as_ref().unwrap().send(UiEvents::ParticipantError(endpoint, err, endpoint_name.clone()));
                             },
                             Message::ParticipantWarning(err) => {
                                 let endpoint_name = self.participants.get_by_right(&endpoint).unwrap();
-                                println!("Participant '{}' Warning: {}", endpoint_name, err);
+                                self.ui_sender.as_ref().unwrap().send(UiEvents::ParticipantWarning(endpoint, err, endpoint_name.clone()));
+                            },
+                            Message::Whisper(err) => {
+                                let endpoint_name = self.participants.get_by_right(&endpoint).unwrap();
+                                self.ui_sender.as_ref().unwrap().send(UiEvents::ParticipantWhisper(endpoint, err, endpoint_name.clone()));
                             }
                             _ => {
+
                                 panic!("Invalid message received by host ({:?})", message);
                             }
                         }
                     }
                     NetEvent::AddedEndpoint(endpoint) => {
-                        println!("Client Added {}", endpoint);
+                        //Client has connected to the host, but at this stage has not yet registered
+                        println!("DEPRECIATED Client Added {}", endpoint);
                     },
                     NetEvent::RemovedEndpoint(endpoint) => {
                         //Client disconnected without unregistering
-                        let endpoint_name = self.participants.get_by_right(&endpoint).unwrap();
-                        println!("Client '{}' has disconnected", endpoint_name);
-
+                        {
+                            let endpoint_name = self.participants.get_by_right(&endpoint).unwrap();
+                            self.ui_sender.as_ref().unwrap().send(UiEvents::ParticipantUnregistered(endpoint, endpoint_name.clone()));
+                        }
                         self.participants.remove_by_right(&endpoint);
                     }
                     NetEvent::DeserializationError(_) => (),
                 },
-                Event::SendCode(code) => {
-                    println!("Sending code to all {} participant(s)", self.participants.len());
+                HostEvent::SendCode(code) => {
+                    //println!("DEPRECIATED Sending code to all {} participant(s)", self.participants.len());
                     for (_name, endpoint) in self.participants.iter() {
                         self.network.send(*endpoint, Message::Code(code.clone()));
                     }
                 },
-                Event::SendData => {
-                    println!("Sending data to all {} participant(s)", self.participants.len());
+                HostEvent::SendData => {
+                    //println!("DEPRECIATED Sending data to all {} participant(s)", self.participants.len());
 
                     //Extract the 'generate data' function from the Lua script.
                     let mut generate_data: hlua::LuaFunction<_> = self.lua.get("generate_data").unwrap();
@@ -241,21 +235,30 @@ impl<'a> Host<'a> {
                         self.network.send(*endpoint, Message::VectorHTP(list));
                     }
                 },
-                Event::Pause(endpoint) => {
+                HostEvent::Pause(endpoint) => {
                     self.network.send(endpoint, Message::Pause);
                 },
-                Event::Play(endpoint) => {
+                HostEvent::Play(endpoint) => {
                     self.network.send(endpoint, Message::Play);
                 },
-                Event::Stop(endpoint) => {
+                HostEvent::Stop(endpoint) => {
                     self.network.send(endpoint, Message::Stop);
                 },
-                Event::Execute => {
-                    println!("Sending execute command to all {} participant(s)", self.participants.len());
-                    for (_name, endpoint) in self.participants.iter() {
+                HostEvent::Execute => {
+                    for (name, endpoint) in self.participants.iter() {
                         self.network.send(*endpoint, Message::Execute);
+                        self.ui_sender.as_ref().unwrap().send(UiEvents::ChangeStatusTo(ParticipantStatus::Calculating, *endpoint, name.clone()))
                     }
                 },
+                HostEvent::Begin(path) => {
+                    self.start_participants(path.as_str());
+                },
+                HostEvent::DebugPrintCount => {
+                    self.display_participant_count();
+                },
+                HostEvent::DebugPrintParticipants => {
+                    self.display_participants();
+                }
             },
             None => {